+++ R/tinyplot.R
@@ -17,7 +17,8 @@
 #'   or interval plot types. Only used when the `type` argument is one of
 #'   `"rect"` or `"segments"` (where all four min-max coordinates are required),
 #'   or `"pointrange"`, `"errorbar"`, or `"ribbon"` (where only `ymin` and
-#'   `ymax` required alongside `x`).
+#'   `ymax` required alongside `x`). In the formula method the arguments
+#'   can be specified as `ymin = var` if `var` is a variable in `data`.
 #' @param by grouping variable(s). The default behaviour is for groups to be
 #'   represented in the form of distinct colours, which will also trigger an
 #'   automatic legend. (See `legend` below for customization options.) However,
@@ -172,15 +173,19 @@
 #' @param xaxs,yaxs character specifying the style of the interval calculation
 #'   used for the x-axis and y-axis, respectively. See
 #'   \code{\link[graphics]{par}} for the possible values.
+#' @param xaxb,yaxb numeric vector (or character vector, if appropriate) giving
+#'   the break points at which the axis tick-marks are to be drawn. Break points
+#'   outside the range of the data will be ignored if the associated axis
+#'   variable is categorical, or an explicit `x/ylim` range is given.
 #' @param xaxl,yaxl a function or a character keyword specifying the format of
 #'   the x- or y-axis tick labels. Note that this is a post-processing step that
-#'   affects the _appearance_ of the tick labels only; it does not affect the
-#'   actual calculation or placement of the tick marks. In addition to
-#'   user-supplied formatting functions (e.g., [`format`], [`toupper`], [`abs`],
-#'   or other custom function), several convenience keywords (or their symbol
-#'   equivalents) are available for common formatting transformations:
-#'   `"percent"` (`"%"`), `"comma"` (`","`), `"log"` (`"l"`), `"dollar"`
-#'   (`"$"`), `"euro"` (`"€"`), or `"sterling"` (`"£"`). See the
+#'   affects the _appearance_ of the tick labels only; use in conjunction with
+#'   `x/yaxb` if you would like to adjust the position of the tick marks too. In
+#'   addition to user-supplied formatting functions (e.g., [`format`],
+#'   [`toupper`], [`abs`], or other custom function), several convenience
+#'   keywords (or their symbol equivalents) are available for common formatting
+#'   transformations: `"percent"` (`"%"`), `"comma"` (`","`), `"log"` (`"l"`),
+#'   `"dollar"` (`"$"`), `"euro"` (`"€"`), or `"sterling"` (`"£"`). See the
 #'   [`tinylabel`] documentation for examples.
 #' @param log a character string which contains `"x"` if the x axis is to be
 #'   logarithmic, `"y"` if the y axis is to be logarithmic and `"xy"` or `"yx"`
@@ -571,6 +576,8 @@ tinyplot.default = function(
     yaxt = NULL,
     xaxs = NULL,
     yaxs = NULL,
+    xaxb = NULL,
+    yaxb = NULL,
     xaxl = NULL,
     yaxl = NULL,
     log = "",
@@ -599,22 +606,23 @@ tinyplot.default = function(
   par_first = get_saved_par("first")
   if (is.null(par_first)) set_saved_par("first", par())
 
+  assert_logical(add)
+  
   # save for tinyplot_add()
-  if (!isTRUE(add)) {
+  if (!add) {
     calls = sys.calls()
     idx = grep("^tinyplot", sapply(calls, function(k) k[[1]]))
     if (length(idx) > 0) {
       options(tinyplot_last_call = calls[[idx[1]]])
     }
+    ## TODO: remove the global option above and move to this when density is refactored
+    # cal = match.call(call = sys.call(sys.parent()), expand.dots = TRUE)
+    # assign(".last_call", cal, envir = get(".tinyplot_env", envir = parent.env(environment())))
   }
-
-  ## TODO: remove the global option above and move to this when density is refactored
-  # cal = match.call(call = sys.call(sys.parent()), expand.dots = TRUE)
-  # assign(".last_call", cal, envir = get(".tinyplot_env", envir = parent.env(environment())))
 
   dots = list(...)
 
-  if (isTRUE(add)) legend = FALSE
+  if (add) legend = FALSE
   draw = substitute(draw)
 
 
@@ -687,7 +695,12 @@ tinyplot.default = function(
   set_saved_par(when = "before", opar)
 
   # catch for adding to existing facet plot
-  if (!is.null(facet) && isTRUE(add)) {
+  if (!is.null(facet) && add) {
+    if (Sys.getenv("POSITRON") == 1) warning(
+      "Positron IDE detected.\n",
+      "Adding layers to a faceted tinyplot in Positron leads to known alignment errors. Please see:\n",
+      "https://github.com/posit-dev/positron/issues/7316"
+    )
     par(get_saved_par(when = "after"))
   }
 
@@ -798,11 +811,13 @@ tinyplot.default = function(
       palette      = palette,
       ribbon.alpha = ribbon.alpha,
       xaxt         = xaxt,
+      xaxb         = xaxb,
       xaxl         = xaxl,
       xlab         = xlab,
       xlabs        = xlabs,
       xlim         = xlim,
       yaxt         = yaxt,
+      yaxb         = yaxb,
       yaxl         = yaxl,
       ylab         = ylab,
       ylim         = ylim
@@ -834,6 +849,9 @@ tinyplot.default = function(
       xaxs_cp = xaxs
       xaxs = yaxs
       yaxs = xaxs_cp
+      xaxb_cp = xaxb
+      xaxb = yaxb
+      yaxb = xaxb_cp
       xaxl_cp = xaxl
       xaxl = yaxl
       yaxl = xaxl_cp
@@ -852,7 +870,7 @@ tinyplot.default = function(
       datapoints[["xmax"]] = if (!is.null(datapoints[["ymax"]])) datapoints[["ymax"]] else NULL
       datapoints[["ymax"]] = if (!is.null(xmax_cp)) xmax_cp else NULL
       # clean up
-      rm(xlim_cp, xlab_cp, xlabs_cp, xaxt_cp, xaxs_cp, xaxl_cp, x_cp, xmin_cp, xmax_cp)
+      rm(xlim_cp, xlab_cp, xlabs_cp, xaxt_cp, xaxs_cp, xaxb_cp, xaxl_cp, x_cp, xmin_cp, xmax_cp)
     } else {
       # We'll let boxplot(..., horizontal = TRUE) handle most of the adjustments
       # and just catch a few elements that we draw beforehand.
@@ -862,10 +880,26 @@ tinyplot.default = function(
       rm(xlab_cp)
     }
   }
-
-
+  
+  # For cases where x/yaxb is provided and corresponding x/ylabs is not null...
+  # We can subset these here to provide breaks
+  if (!is.null(xaxb) && !is.null(xlabs)) {
+    xlabs = xlabs[names(xlabs) %in% xaxb]
+    xaxb = NULL # don't need this any more
+  }
+  if (!is.null(yaxb) && !is.null(ylabs)) {
+    ylabs = ylabs[names(ylabs) %in% yaxb]
+    yaxb = NULL # don't need this any more
+  }
+  
   # plot limits
-  fargs = lim_args(datapoints = datapoints, xlim = xlim, ylim = ylim, type = type)
+  fargs = lim_args(
+    datapoints = datapoints,
+    xlim = xlim, ylim = ylim,
+    xaxb = xaxb, yaxb = yaxb,
+    xlim_user = xlim_user, ylim_user = ylim_user,
+    type = type
+  )
   list2env(fargs, environment())
 
 
@@ -961,7 +995,7 @@ tinyplot.default = function(
     }
   }
 
-  if ((is.null(legend) || legend != "none") && isFALSE(add)) {
+  if ((is.null(legend) || legend != "none") && !add) {
     if (isFALSE(by_continuous)) {
       if (ngrps > 1) {
         lgnd_labs = if (is.factor(datapoints$by)) levels(datapoints$by) else unique(datapoints$by)
@@ -994,7 +1028,7 @@ tinyplot.default = function(
     )
 
     has_legend = TRUE
-  } else if (legend_args[["x"]] == "none" && isFALSE(add)) {
+  } else if (legend_args[["x"]] == "none" && !add) {
     omar = par("mar")
     ooma = par("oma")
     topmar_epsilon = 0.1
@@ -1024,7 +1058,7 @@ tinyplot.default = function(
   }
 
   # Titles. Only draw these if add = FALSE
-  if (isFALSE(add)) {
+  if (!add) {
     # main title
     # Note that we include a special catch for the main title if legend is
     # "top!" (and main is specified in the first place).
@@ -1105,7 +1139,7 @@ tinyplot.default = function(
   # placeholders for facet_window_args() call
   facet_newlines = facet_text = facet_rect = facet_font = facet_col = facet_bg = facet_border = NULL
 
-  if (!is.null(facet) && isFALSE(add)) {
+  if (!is.null(facet) && !add) {
     if (is.null(omar)) omar = par("mar")
 
     # Grab some of the customizable facet args that we'll be using later
@@ -1154,8 +1188,8 @@ tinyplot.default = function(
       # axes args
       axes = axes, flip = flip, frame.plot = frame.plot,
       oxaxis = oxaxis, oyaxis = oyaxis,
-      xlabs = xlabs, xlim = xlim, xlim_user = xlim_user, xaxt = xaxt, xaxs = xaxs, xaxl = xaxl,
-      ylabs = ylabs, ylim = ylim, ylim_user = ylim_user, yaxt = yaxt, yaxs = yaxs, yaxl = yaxl,
+      xlabs = xlabs, xlim = xlim, xlim_user = xlim_user, xaxt = xaxt, xaxs = xaxs, xaxb = xaxb, xaxl = xaxl,
+      ylabs = ylabs, ylim = ylim, ylim_user = ylim_user, yaxt = yaxt, yaxs = yaxs, yaxb = yaxb, yaxl = yaxl,
       asp = asp, log = log,
       # other args (in approx. alphabetical + group ordering)
       dots = dots,
@@ -1178,8 +1212,8 @@ tinyplot.default = function(
       nfacets = nfacets, nfacet_cols = nfacet_cols, nfacet_rows = nfacet_rows,
       axes = axes, flip = flip, frame.plot = frame.plot,
       oxaxis = oxaxis, oyaxis = oyaxis,
-      xlabs = xlabs, xlim = xlim, xlim_user = xlim_user, xaxt = xaxt, xaxs = xaxs, xaxl = xaxl,
-      ylabs = ylabs, ylim = ylim, ylim_user = ylim_user, yaxt = yaxt, yaxs = yaxs, yaxl = yaxl,
+      xlabs = xlabs, xlim = xlim, xlim_user = xlim_user, xaxt = xaxt, xaxs = xaxs, xaxb = xaxb, xaxl = xaxl,
+      ylabs = ylabs, ylim = ylim, ylim_user = ylim_user, yaxt = yaxt, yaxs = yaxs, yaxb = yaxb, yaxl = yaxl,
       asp = asp, log = log,
       dots = dots,
       draw = draw,
@@ -1328,9 +1362,11 @@ tinyplot.default = function(
     }
   }
   
-  # save end pars for possible recall later
-  apar = par(no.readonly = TRUE)
-  set_saved_par(when = "after", apar)
+  if (!add) {
+    # save end pars for possible recall later
+    apar = par(no.readonly = TRUE)
+    set_saved_par(when = "after", apar)
+  }
 
 }
 
@@ -1346,6 +1382,10 @@ tinyplot.formula = function(
     facet = NULL,
     facet.args = NULL,
     type = NULL,
+    xmin = NULL,
+    xmax = NULL,
+    ymin = NULL,
+    ymax = NULL,
     xlim = NULL,
     ylim = NULL,
     # log = "",
@@ -1384,12 +1424,18 @@ tinyplot.formula = function(
   ## placeholder for legend title
   legend_args = list(x = NULL)
 
+  ## turn facet into a formula if it does not evaluate successfully
+  if (inherits(try(facet, silent = TRUE), "try-error")) {
+    facet = as.formula(paste("~", deparse(substitute(facet))))
+    environment(facet) = environment(formula)
+  }
+
   ## process all formulas
   tf = tinyformula(formula, facet)
 
   ## set up model frame
   m = match.call(expand.dots = FALSE)
-  m = m[c(1L, match(c("formula", "data", "subset", "na.action", "drop.unused.levels"), names(m), 0L))]
+  m = m[c(1L, match(c("formula", "data", "subset", "na.action", "drop.unused.levels", "xmin", "xmax", "ymin", "ymax"), names(m), 0L))]
   m$formula = tf$full
   ## need stats:: for non-standard evaluation
   m[[1L]] = quote(stats::model.frame)
@@ -1459,6 +1505,10 @@ tinyplot.formula = function(
     facet = facet, facet.args = facet.args,
     data = data,
     type = type,
+    xmin = mf[["(xmin)"]],
+    xmax = mf[["(xmax)"]],
+    ymin = mf[["(ymin)"]],
+    ymax = mf[["(ymax)"]],
     xlim = xlim,
     ylim = ylim,
     # log = "",